utci plot methods fixe
