+++ LadybugTools_Engine/Python/src/ladybugtools_toolkit/plot/colormaps.py
@@ -1,0 +1,12 @@
+from .utilities import colormap_sequential
+
+DBT_COLORMAP = colormap_sequential("white", "#bc204b")
+RH_COLORMAP = colormap_sequential("white", "#8db9ca")
+MRT_COLORMAP = colormap_sequential("white", "#6d104e")
+WS_COLORMAP = colormap_sequential(
+    "#d0e8e4",
+    "#8db9ca",
+    "#006da8",
+    "#24135f",
+)
+UTCI_DIFFERENCE_COLORMAP = colormap_sequential("#00A9E0", "w", "#702F8A")
